requirements.txt
opencv-python
ultralytics
google-generativeai
yt-dlp
python-dotenv
matplotlib
numpy